gita-bot/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
langchain-community==0.3.0
transformers==4.44.2
PyPDF2==3.0.1
python-dotenv==1.0.1
torch==2.4.1
sentence-transformers==3.1.1